{"types":"C","runs":15}

+++ Problem15.java
@@ -4,7 +4,7 @@ import acm.program.ConsoleProgram;
 public class Problem15 extends ConsoleProgram {
 	public void run(){
 		double num1 = readInt("Write first whole number ");
-		double num2 = readInt("Write second whole number ");
+		int num2 = readInt("Write second whole number ");
 		double avg = (num1+num2)/2;
 		println("Average of this two numbers is "+ avg);
